+++ noodles-fasta/CHANGELOG.md
@@ -6,11 +6,17 @@
 
   * fasta/fai: Move reader (`Reader`) and writer (`Writer`) to `io` module.
 
+  * fasta/fai/async: Move reader (`Reader`) to `io` module.
+
 ### Deprecated
 
   * fasta/fai: Deprecate `Reader` and `Writer`.
 
     Use `fai::io::Reader` and `fai::io::Writer`, respectively, instead.
+
+  * fasta/fai/async: Deprecate `Reader` and `Writer`.
+
+    Use `fai::r#async::io::Reader` instead.
 
 ## 0.43.0 - 2024-09-04